Add some comments and misc changes
